+++ src/sas/qtgui/Perspectives/Corfunc/corefuncutil.py
@@ -20,14 +20,3 @@ WIDGETS = enum( 'W_QMIN',
                 'W_PERIOD',
                 'W_FILENAME',
                 )
-
-@dataclass
-class ExtractedParameters:
-    long_period: float
-    interface_thickness: float
-    hard_block_thickness: float
-    soft_block_thickness: float
-    core_thickness: float
-    polydispersity_ryan: float
-    polydispersity_stribeck: float
-    local_crystallinity: float
