+++ test/corfunc/utest_corfunc.py
@@ -88,11 +88,7 @@ class TestCalculator(unittest.TestCase):
     def extract_params(self):
         params = self.calculator.extract_parameters(self.transformation[0])
 
-        # Removing the tests below, will be scrapping the dict for a dataclass in dependant PR very soon
-        #self.assertIsNotNone(params)
-        #self.assertEqual(len(params), 6)
-
-        self.assertLess(abs(params['max']-75), 2.5) # L_p ~= 75
+        self.assertLess(abs(params.long_period-75), 2.5) # L_p ~= 75
 
     def check_transforms(self):
         gamma1, gamma3, idf = self.transformation
